+++ types/dto/AeroflyMission.d.ts
@@ -202,5 +202,6 @@ export declare class AeroflyMission {
      * @returns {string} to use in Aerofly FS4's `custom_missions_user.tmc`
      */
     toString(): string;
+    static fromJSON(json: unknown): AeroflyMission;
 }
 //# sourceMappingURL=AeroflyMission.d.ts.map